pipeline/pyproject.toml
[project]
name = "lennys-polls-pipeline"
version = "0.1.0"
description = "Automated survey-to-dashboard pipeline for Lenny's Polls"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "anthropic>=0.40",
    "jinja2>=3.1",
    "pandas>=2.2",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.scripts]
pipeline = "pipeline.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
